# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benin Tourisme API",
        "description": "API de gestion et promotion des sites touristiques et événements culturels du Bénin",
        "contact": {
            "email": "eldomoreogbohouili@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API local"
        }
    ],
    "paths": {
        "/api/admin/admins": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Lister les administrateurs",
                "operationId": "9a660a06a7f3fd979fa2898bf16ae873",
                "responses": {
                    "200": {
                        "description": "Liste des admins"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Créer un admin",
                "operationId": "b7126ca6727f1f59ffb8bcf2b9adc57c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "tel",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin créé"
                    }
                }
            }
        },
        "/api/admin/admins/{id}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Afficher un admin",
                "operationId": "0c78ef17cd24543e448aec8670840ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Mettre à jour un admin",
                "operationId": "2326fae561c671672396afe5b0ec2b00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Supprimer un admin",
                "operationId": "58672fc8d14eb709b76e9c5650ffd64f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}